Fixed data type for GetSpeciesHeight/Weight

+++ src/pokemon.c
@@ -6267,12 +6267,12 @@ const u8 *GetSpeciesPokedexDescription(u16 species)
     return gSpeciesInfo[species].description;
 }
 
-u8 GetSpeciesHeight(u16 species)
+u16 GetSpeciesHeight(u16 species)
 {
     return gSpeciesInfo[SanitizeSpeciesId(species)].height;
 }
 
-u8 GetSpeciesWeight(u16 species)
+u16 GetSpeciesWeight(u16 species)
 {
     return gSpeciesInfo[SanitizeSpeciesId(species)].weight;
 }
